+++ gsa/src/gmp/parser/cvss.js
